+++ src/camera.cpp
@@ -7,7 +7,7 @@
 
 #define PI 3.14159265359
 
-Camera::Camera(const Vector3d &pos, const Vector3d &dir, const Vector3d &up, const double fov, const double aspect_ratio,
+Camera::Camera(const Vector3f &pos, const Vector3f &dir, const Vector3f &up, const double fov, const double aspect_ratio,
 		const Vector2i &resolution, double focal_length, double blur_radius)
 		: m_position(pos), m_direction(dir), m_up(up), m_blur_radius(blur_radius), m_fov(fov),
 			m_aspect_ratio(aspect_ratio), m_focal_length(focal_length),	m_resolution(resolution) 
@@ -23,7 +23,7 @@ void Camera::update()
 	double image_plane_height = 2 * distance * tan(m_fov / 2.0);
 	double image_plane_width = image_plane_height * m_aspect_ratio;
     
-    Vector3d n = m_direction * -1;
+    Vector3f n = m_direction * -1;
     n.normalize();
 
     m_u = m_up.cross(n);
@@ -32,7 +32,7 @@ void Camera::update()
 	m_v = n.cross(m_u);
 	m_v.normalize();
 
-	Vector3d image_plane_center = m_position - n * distance;
+	Vector3f image_plane_center = m_position - n * distance;
 
 	m_image_plane_origin = image_plane_center + (m_u * (image_plane_width / 2.0)) - (m_v * (image_plane_height / 2.0));
 
@@ -53,7 +53,7 @@ void Camera::cast_perturbed_ray(Ray &ray, size_t x, size_t y, std::mt19937 &rand
 {
 	cast_ray(ray, x, y);
 
-	Vector3d focus_point = m_position + ray.m_direction * m_focal_length;
+	Vector3f focus_point = m_position + ray.m_direction * m_focal_length;
 
 	std::uniform_real_distribution<> distribution(0, 1);
