+++ src/ray.h
@@ -11,7 +11,7 @@ public:
 	{
 		dist = DBL_MAX;
 	}
-	Ray(const Vector3d &origin, const Vector3d &direction)
+	Ray(const Vector3f &origin, const Vector3f &direction)
 		: m_origin(origin), m_direction(direction)
 	{
 		m_dir_reciprocal.x() = 1.0 / m_direction.x();
@@ -21,13 +21,13 @@ public:
 		dist = DBL_MAX;
 	}
 
-	Vector3d m_origin;
-	Vector3d m_direction;
-	Vector3d m_dir_reciprocal;
+	Vector3f m_origin;
+	Vector3f m_direction;
+	Vector3f m_dir_reciprocal;
 
 	double dist;
 
-	bool intersects(const Vector3d &v1, const Vector3d &v2, const Vector3d &v3, double *u, double *v, double *dist) const;
+	bool intersects(const Vector3f &v1, const Vector3f &v2, const Vector3f &v3, double *u, double *v, double *dist) const;
 };
 
 #endif
